+++ crypt-core/src/main/java/de/alpharogroup/crypto/blockchain/Block.java
@@ -52,15 +52,15 @@ public class Block implements IBlock
 
 	byte[] hash;
 
+	byte[] merkleRoot;
+
 	byte[] previousBlockHash;
+
+	long timestamp;
 
 	List<ITransaction> transactions;
 
-	byte[] merkleRoot;
-
 	long tries;
-
-	long timestamp;
 
 	public Block(byte[] previousBlockHash, List<ITransaction> transactions, long tries)
 	{
@@ -68,8 +68,9 @@ public class Block implements IBlock
 		this.transactions = transactions;
 		this.tries = tries;
 		this.timestamp = System.currentTimeMillis();
-		this.merkleRoot = HashExtensions.getMerkleRootHash(new LinkedList<>(
-			transactions.stream().map(ITransaction::getHash).collect(Collectors.toList())),
+		this.merkleRoot = HashExtensions.getMerkleRootHash(
+			new LinkedList<>(
+				transactions.stream().map(ITransaction::getHash).collect(Collectors.toList())),
 			HashAlgorithm.SHA256);
 		this.hash = HashExtensions.hash(previousBlockHash, merkleRoot, tries, timestamp,
 			HashAlgorithm.SHA256);